Welcome to ManoptExample.jl

ManoptExamples.ManoptExamples — Module

🏔️⛷️ ManoptExamples.jl – A collection of research and tutorial example problems for Manopt.jl

📚 Documentation: juliamanifolds.github.io/ManoptExamples.jl

📦 Repository: github.com/JuliaManifolds/ManoptExamples.jl

💬 Discussions: github.com/JuliaManifolds/ManoptExamples.jl/discussions

🎯 Issues: github.com/JuliaManifolds/ManoptExamples.jl/issues

This package provides a set of example tasks for Manopt.jl based on either generic manifolds from the ManifoldsBase.jl interface or specific manifolds from Manifolds.jl.

Each example usually consists of

a cost function and additional objects, like the gradient or proximal maps, see objectives

an example explaining how to use these, see examples

Helping functions that are used in one or more examples can be found in the section of functions in the menu.